Prefer working with int colors and ObjectList.
ArrayList is fine, but ObjectList extends it in some useful ways, and interops with the rest of jdkgdxds more thoroughly. It can also be written by jdkgdxds-interop. LinkedList is usually less-fine. Using int colors is how SquidSquad prefers to work now, because the mutable libGDX Color is just a mess, and primitives make sense for frequently-created values.