Update SquidSquad, migrate some Lists to Deques.
DijkstraMap uses ObjectDeque for its paths now, because those paths were just as frequently altered from the start as from the end of the ordering, and ObjectList (like ArrayList) can be inefficient when editing the start of a List. Unlike ArrayDeque, ObjectDeque actually allows random-access lookup in constant time, so it can be used in many places ArrayList or ObjectList would.